Move AMQP worker connection utility functions to new amqp module
