+++ src/actions/status.rs
@@ -6,7 +6,10 @@ use std::sync::mpsc::{Receiver, TryRecvError};
 use std::sync::{mpsc, Arc, Mutex};
 use std::time::Duration;
 
-use clap::{Arg, ArgMatches};
+use crate::amqp::{
+  get_amqp_server_url, get_request_headers, DIRECT_MESSAGING, DIRECT_MESSAGING_RESPONSE,
+};
+use clap::ArgMatches;
 use crossterm::terminal::{Clear, ClearType};
 use crossterm::{cursor, QueueableCommand};
 use futures_util::stream::StreamExt;
@@ -14,7 +17,7 @@ use lapin::options::BasicAckOptions;
 use lapin::publisher_confirm::PublisherConfirm;
 use lapin::{
   options::{BasicConsumeOptions, BasicPublishOptions},
-  types::{AMQPValue, FieldTable},
+  types::FieldTable,
   uri::AMQPUri,
   BasicProperties, Connection, ConnectionProperties,
 };
@@ -25,44 +28,6 @@ use mcai_worker_sdk::{
   worker::system_information::SystemInformation,
   Channel,
 };
-
-const DIRECT_MESSAGING: &str = "direct_messaging";
-const DIRECT_MESSAGING_RESPONSE: &str = "direct_messaging_response";
-
-pub fn get_command_args() -> [Arg<'static, 'static>; 7] {
-  [
-    Arg::with_name("host")
-      .short("h")
-      .long("host")
-      .takes_value(true)
-      .default_value("localhost"),
-    Arg::with_name("port")
-      .short("p")
-      .long("port")
-      .takes_value(true)
-      .default_value("5672"),
-    Arg::with_name("virtual_host")
-      .short("vh")
-      .long("vhost")
-      .takes_value(true)
-      .default_value(""),
-    Arg::with_name("user")
-      .short("u")
-      .long("user")
-      .takes_value(true)
-      .default_value("guest"),
-    Arg::with_name("password")
-      .short("P")
-      .long("password")
-      .takes_value(true)
-      .default_value("guest"),
-    Arg::with_name("tls").long("tls"),
-    Arg::with_name("worker_id")
-      .short("w")
-      .long("worker-id")
-      .takes_value(true),
-  ]
-}
 
 pub fn status(matches: &ArgMatches) {
   match get_amqp_server_url(matches) {
@@ -170,27 +135,6 @@ pub fn get_worker_status(
   Ok(())
 }
 
-fn get_amqp_server_url(matches: &ArgMatches) -> Result<String, String> {
-  let scheme = if matches.is_present("tls") {
-    "amqps"
-  } else {
-    "amqp"
-  };
-
-  let host = matches.value_of("host").unwrap();
-  let port = matches.value_of("port").unwrap();
-  let user = matches.value_of("user").unwrap();
-  let password = matches.value_of("password").unwrap();
-  let virtual_host = matches.value_of("virtual_host").unwrap();
-
-  let server_url = format!(
-    "{}://{}:{}@{}:{}/{}",
-    scheme, user, password, host, port, virtual_host
-  );
-
-  Ok(server_url)
-}
-
 fn declare_consumed_queue(channel: &Channel) {
   let direct_messaging_response_queue = QueueDescription {
     name: DIRECT_MESSAGING_RESPONSE.to_string(),
@@ -225,19 +169,6 @@ fn send_status_request(channel: &Channel, headers: FieldTable) -> Result<Publish
     )
     .wait()
     .map_err(|e| e.to_string())
-}
-
-fn get_request_headers(worker_id: Option<&str>) -> FieldTable {
-  let mut headers = FieldTable::default();
-  if let Some(worker) = worker_id {
-    headers.insert(
-      "worker_name".into(),
-      AMQPValue::LongString(worker.to_string().into()),
-    );
-  } else {
-    headers.insert("broadcast".into(), AMQPValue::Boolean(true));
-  }
-  headers
 }
 
 fn stop_consumer(rx: &Receiver<()>) -> bool {
